chore: change ConnectFederation to ConnectFederationModal
