+++ apps/gateway-ui/src/components/index.tsx
@@ -1,7 +1,7 @@
 export { GatewayCard } from './GatewayCard';
 export { WithdrawCard } from './WithdrawCard';
 export { FederationCard } from './FederationCard';
-export { ConnectFederation } from './ConnectFederation';
+export { ConnectFederationModal } from './ConnectFederationModal';
 export { InfoCard } from './InfoCard';
 export { DepositCard } from './DepositCard';
 export { BalanceCard } from './BalanceCard';
